refactor(tests): add unit tests for response normalizers
